requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
pyaes==1.6.1
pyasn1==0.4.8
PySocks==1.7.1
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rsa==4.8
six==1.16.0
Telethon==1.24.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
